refactor: close GF11 by moving spec dispatch ownership
Promote SpecCommands and execute_command_value into the shared spec module, keep commands/spec as a CLI adapter, and route Mother builtin spec dispatch through patina::spec so the greenfield lane reaches full GF1-GF12 closure.
