+++ src/commands/spec/internal/mod.rs
@@ -41,5 +41,6 @@ pub(crate) use create::create_spec_value;
 // Split _value() function re-exported pub(crate) by parent for MCP
 pub(crate) use split::split_spec_value;
 
-// Functions called from spec/mod.rs execute path
-pub(super) use archive::{archive_spec, archive_stale_specs};
+#[allow(unused_imports)]
+// Functions called from spec execute path
+pub(crate) use archive::{archive_spec, archive_stale_specs};
